Kappa 4 support, expression improvements (#334)
* Kappa 4 support, Kappa expression improvements

This PR introduces support for Kappa 4 and drops support for
Kappa 3. Note that Kappa now requires agents on the reactant
side to correspond to agents on the product side of a rule
(i.e. order is preserved). The "dot" agent is implemented in
PySB as `None`, and is necessary in
[certain situations documented in the Kappa Manual](https://github.com/Kappa-Dev/KaSim/blob/master/CHANGES.md#agent-order-in-rules-must-be-maintained).

For example, the following Kappa rule:

    // Agent A spawns a copy of agent B
    A(),. -> A(),B() @ 'r'

would be implemented in PySB as:

    Rule('spawnB', A() + None >> A() + B(), r)

This PR also improves Expression support for Kappa, implementing
`log`, `exp`, `min`, `max`, `sin`, `cos`, `tan`, `mod` and `sqrt`
functions, if expressions via sympy's `Piecewise`, and the `pi` and
`E` constants.

When using BioNetGen, `None` in a reaction pattern is translated
to the null symbol `0`.

Fixes: #333
Fixes: #288